🚚 [templates] Rename module `{binders => bindvariables}`
