+++ src/cutty/projects/generate.py
@@ -4,7 +4,7 @@ from collections.abc import Sequence
 from cutty.projects.generator import ProjectGenerator
 from cutty.projects.project import Project
 from cutty.projects.template import Template
-from cutty.templates.domain.binders import bindvariables
+from cutty.templates.domain.bindvariables import bindvariables
 from cutty.variables.bindings import Binding
 from cutty.variables.prompts import createprompt
